refactor(runtime-coordination): decompose 1164→68L facade with 3 modules (F27 PR-2)
Extract runtime_coordination.py (1164L) into:
- coordination_db.py (275L): state constants, exceptions, DB connection, schema, queries
- runtime_state_machine.py (284L): dispatch state validation, transitions, attempts
- lease_operations.py (318L): raw lease lifecycle functions with BOOT-9/10/11 decomposition
- runtime_coordination.py (68L): backward-compatible re-export facade

All existing `from runtime_coordination import X` continue to work unchanged.
No circular imports: coordination_db is leaf, others import from it directly.
lease_manager.py (419L, LeaseManager class) unchanged — imports via facade.

Dispatch-ID: 20260405-120020-f27-pr2-runtime-coord-retry-C
